+++ BadgeLedApp/Localizable.xcstrings
@@ -75,6 +75,9 @@
     },
     "Speed:" : {
 
+    },
+    "Weight" : {
+
     }
   },
   "version" : "1.0"
