+++ BadgeLedApp/PixelGrid/View/PixelGridView.swift
@@ -1,66 +1,39 @@
 //
-//  PixelGridImage.swift
+//  PixelEditorView 2.swift
 //  BadgeLedApp
 //
 //  Created by Gustav on 30.12.24.
 //
 import SwiftUI
 
-struct PixelGridImage: View {
-    let width: Int
-    let pixels: [[Pixel]]
+struct PixelGridView: View {
+    @ObservedObject var pixelGrid: PixelGrid
+    @State private var drawMode: Bool = true
+    @Environment(\.undoManager) var undo
     
     var body: some View {
-        Canvas { context, size in
-            let pixelWidth: CGFloat = size.width / CGFloat(width)
-            let pixelHeight: CGFloat = size.height / 11
-            
-            // Create symbol once and reuse
-            guard let itemOn = context.resolveSymbol(id: "itemOn") else { return }
-            guard let itemOff = context.resolveSymbol(id: "itemOff") else { return }
-            
-            let onPixels: [Pixel] = pixels.compactMap { row in
-                row.compactMap { pixel in
-                    pixel.isOn ? pixel : nil
-                }
-            }.flatMap { $0 }
-            
-            for pixel in onPixels {
-                context.draw(
-                    itemOn,
-                    at: CGPoint(
-                        x: CGFloat(pixel.x) * pixelWidth + pixelWidth/2,
-                        y: CGFloat(pixel.y) * pixelHeight + pixelHeight/2
-                    )
+        VStack {
+            PixelGridImage(pixelGrid: pixelGrid)
+                .frame(width: CGFloat(pixelGrid.width * 20),
+                       height: CGFloat(11 * 20))
+                .gesture(
+                    DragGesture(minimumDistance: 0)
+                        .onChanged { value in
+                            let pixelSize: CGFloat = 20 // 20px + 1px spacing
+                            
+                            let x = Int(value.location.x / pixelSize)
+                            let y = Int(value.location.y / pixelSize)
+                            
+                            if x >= 0 && x < pixelGrid.width && y >= 0 && y < pixelGrid.height {
+                                if value.translation == .zero {
+                                    // This is the start of the drag - set mode based on initial pixel
+                                    drawMode = !pixelGrid.pixels[y][x].isOn
+                                }
+                                pixelGrid.setPixel(x: x, y: y, isOn: drawMode, undoManager: undo)
+                            }
+                        }
                 )
-            }
             
-            let offPixels: [Pixel] = pixels.compactMap { row in
-                row.compactMap { pixel in
-                    pixel.isOn == false ? pixel : nil
-                }
-            }.flatMap { $0 }
-            
-            for pixel in offPixels {
-                context.draw(
-                    itemOff,
-                    at: CGPoint(
-                        x: CGFloat(pixel.x) * pixelWidth + pixelWidth/2,
-                        y: CGFloat(pixel.y) * pixelHeight + pixelHeight/2
-                    )
-                )
-            }
-            
-        } symbols: {
-            RoundedRectangle(cornerRadius: 3, style: .continuous)
-                .fill(Color.accentColor)
-                .frame(width: 19, height: 19)
-                .tag("itemOn")
-            
-            RoundedRectangle(cornerRadius: 3, style: .continuous)
-                .fill(.tertiary)
-                .frame(width: 19, height: 19)
-                .tag("itemOff")
         }
     }
 }
